requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
motor  # async Mongo client
python-dotenv
python-dateutil
pymongo

